EKS/requirements.txt
flask
kubernetes
pyyaml
opencv-contrib-python
bottle
tqdm
pymongo